remote_container/terraform/cloud_function_source/requirements.txt
functions-framework==3.*
qrcode==7.4.2
pillow==10.3.0
google-cloud-storage==2.17.0